% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexGrid.R
\name{climdexGrid}
\alias{climdexGrid}
\title{ETCCDI Climate Changes Indices in Climate4R}
\usage{
climdexGrid(index.code, tn = NULL, tx = NULL, pr = NULL,
  input.arg.list = list(), index.arg.list = list(), cal = "365_day",
  parallel = FALSE, max.ncores = 16, ncores = NULL)
}
\arguments{
\item{index.code}{Character string, indicating the specific code of the index according to the ETCCDI
definitions (see Details).}

\item{tn}{A climate4R dataset of daily minimum temperature (degrees C)}

\item{tx}{A climate4R dataset of daily maximum temperature (degrees C)}

\item{pr}{A climate4R dataset of daily precipitation (mm)}

\item{input.arg.list}{Optional. A list of arguments internally passed to \code{\link[climdex.pcic]{climdexInput.raw}}
from package \pkg{climdex.pcic}}

\item{index.arg.list}{Optional (but depending on the specific index might be necessary). A list of specific arguments
for the target index. See Details.}

\item{cal}{A calendar definition. Default to 365-day calendar. This argument is passed to \code{\link[PCICt]{as.PCICt}},
whose help documentation contains further details.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\description{
Calculation of the 27 core indices of the Expert Team on Climate Change Detection and Indices (ETCCDI).
The function is a wrapper of the package \pkg{climdex.pcic} for its seamless integration with climate4R objects.
}
\details{
\code{\link{climdexShow}} will display on screen a full list of ETCCDI Core indices and their codes. The names of the
internal functions calculating each index is also displayed, whose help files can aid in the definition of index-specific arguments.

\strong{Baseline period}
By default, the function will use as baseline period the full period of years encompassed by the input grid(s). This
is used, for instance, for calculating the relevant percentiles in some indices etc. To use a specific baseline period
use the \code{"base.range"} argument in the \code{input.arg.list} internally passed to
 \code{\link[climdex.pcic]{climdexInput.raw}}.
}
\section{Parallel Processing}{


Parallel processing is enabled using the \pkg{parallel} package. 
Parallelization is undertaken by a FORK-type parallel socket cluster formed by \code{ncores}.
If \code{ncores} is not specified (default), \code{ncores} will be one less than the autodetected number of cores.
The maximum number of cores used for parallel processing can be set with the \code{max.ncores} argument, 
although this will be reset to the auto-detected number of cores minus 1 if this number is exceeded. Note that not all 
code, but just some critical loops within the function are parallelized.

In practice, parallelization does not always result in smaller execution times, due to the parallel overhead.
However, parallel computing may potentially provide a significant speedup for the 
particular case of large multimember datasets or large grids.
 
Parallel computing is currently not available for Windows machines.
}

\examples{
\dontrun{
require(climate4R.climdex)
require(visualizeR)
data("tasmin.eobs")
## FROST DAYS (Annual count of days when TN < 0 degC)
fd.grid <- climdexGrid(tn = tasmin.eobs, index.code = "FD")
spatialPlot(climatology(fd.grid), at = seq(0,165,10),
            main = "Mean annual number of frost days (1991-2010)")


# The following example will compute the CWD index
# (maximum number of consecutive days with RR ≥ 1mm)
# for the homogeneized VALUE dataset of stations over Europe
# (See Gutiérrez et al. 2018 DOI:10.1002/joc.5462 for details on this dataset)
library(loadeR)
library(transformeR)
library(visualizeR)
destfile = "/tmp/VALUE_ECA_86_v2.tar.gz"
# (~8Mb download, change destfile at your convenience)
download.file("http://meteo.unican.es/work/loadeR/data/VALUE_ECA_86_v2.tar.gz", destfile = destfile)
untar(destfile, exdir = "/tmp")
station.data <- loadStationData(dataset = "/tmp/VALUE_ECA_86_v2",
                                var = "precip",
                                years = 1981:2000)
cwd <- climdexGrid(index.code = "CWD", pr = station.data)
spatialPlot(climatology(cwd), backdrop.theme = "countries",
            main = "Mean number of consecutive annual wet days (1981-2000)")
}

}
\author{
J. Bedia
}
