% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexGrid.R
\name{climdexShow}
\alias{climdexShow}
\title{List all the 27 ETCCDI Core Indices}
\usage{
climdexShow()
}
\value{
Print a table on the screen with the following columns:
\itemize{
\item \strong{code}: Code of the index. This is the character string used as input value
for the argument \code{index.code} in \code{\link{climdexGrid}}
\item \strong{longname}: Long description of the index
\item \strong{index.fun}: The name of the internal function from package \pkg{\link{climdex.pcic}} used to calculate it
\item \strong{tn,tx,pr}: A logical value (0/1) indicating the input variables required for index calculation
\item \strong{units}: The units of the index (when different from those of the input variable)
}
}
\description{
Print a table with a summary of the 27 ETCCDI Core indices
}
\references{
The ETCDDI web page giving the definition of the 27 core indices:
\url{http://etccdi.pacificclimate.org/list_27_indices.shtml}
}
\author{
J. Bedia
}
