% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdexGrid.R
\name{climdexGrid}
\alias{climdexGrid}
\title{ETCCDI Climate Changes Indices in Climate4R}
\usage{
climdexGrid(tn = NULL, tx = NULL, pr = NULL, index.code,
  input.arg.list = list(), index.arg.list = list(), cal = "365_day",
  parallel = FALSE, max.ncores = 16, ncores = NULL)
}
\arguments{
\item{tn}{A climate4R grid of daily minimum temperature}

\item{tx}{A climate4R grid of daily maximum temperature}

\item{pr}{A climate4R grid of daily precipitation}

\item{index.code}{Character string, indicating the specific code of the index according to the ETCCDI
definitions (see Details).}

\item{input.arg.list}{Optional. A list of arguments internally passed to \code{\link[climdex.pcic]{climdexInput.raw}}
from package \pkg{climdex.pcic}}

\item{index.arg.list}{Optional (but depending on the specific index might be necessary). A list of specific arguments
for the target index. See Details.}

\item{cal}{A calendar definition. Default to 365-day calendar. This argument is passed to \code{\link[PCICt]{as.PCICt}},
whose help documentation contains further details.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\description{
Calculation of the 27 core indices of the Expert Team on Climate Change Detection and Indices (ETCCDI).
The function is a wrapper of the package \pkg{climdex.pcic} for its seamless integration with climate4R objects.
}
\details{
\code{\link{climdexShow}} will display on screen a full list of ETCCDI Core indices and their codes. The names of the
internal functions calculating each index is also displayed, whose help files can aid in the definition of index-specific arguments.

\strong{Parallel computing}
}
\section{Parallel Processing}{


Parallel processing is enabled using the \pkg{parallel} package. 
Parallelization is undertaken by a FORK-type parallel socket cluster formed by \code{ncores}.
If \code{ncores} is not specified (default), \code{ncores} will be one less than the autodetected number of cores.
The maximum number of cores used for parallel processing can be set with the \code{max.ncores} argument, 
although this will be reset to the auto-detected number of cores minus 1 if this number is exceeded. Note that not all 
code, but just some critical loops within the function are parallelized.

In practice, parallelization does not always result in smaller execution times, due to the parallel overhead.
However, parallel computing may potentially provide a significant speedup for the 
particular case of large multimember datasets or large grids.
 
Parallel computing is currently not available for Windows machines.
}

\author{
J. Bedia
}
